chore: split implementations into multiple files for maintainability
