+++ internal/api/pkg_helm.go
@@ -8,40 +8,9 @@ import (
 	"net/http"
 	openapi_types "github.com/oapi-codegen/runtime/types"
 
-	"artifacts/internal/storage"
+	//"artifacts/internal/storage"
 	"artifacts/internal/storage/storage_error"
 )
-
-var chartMIMEType = []string{"application/gzip"}
-
-type Server struct{
-	storageHandler storage.Storage
-}
-
-func NewServer(storageHandler storage.Storage) Server {
-	return Server{
-		storageHandler: storageHandler,
-	}
-}
-
-func (s Server) GetRepositories(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	_ = json.NewEncoder(w).Encode(Error{
-		Code: new(int(http.StatusOK)),
-		Message: new(string("")),
-	})
-	return
-}
-
-func (s Server) AddRepository(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	_ = json.NewEncoder(w).Encode(Error{
-		Code: new(int(http.StatusCreated)),
-		Message: new(string("Repository created")),
-	})
-	return
-}
 
 func (s Server) GetCharts(w http.ResponseWriter, r *http.Request) {
 	data, err := s.storageHandler.Read("", "")
